src/indexer/pyproject.toml
[tool.poetry]
name = "maven-indexer"
version = "0.1.0"
description = ""
authors = ["Tristan Allaire <tristan.allaire@alarch.fr>"]

[tool.poetry.dependencies]
python = "<3.13,>=3.12"
dipdup = "8.0.0"
python-dateutil = "^2.9.0.post0"
PyYAML = "^6.0"
asyncclick = "^8.1.3"
python-telegram-bot = "^20.2"
tortoise-orm = "^0.21.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
